backend/requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.3
psycopg2-binary==2.9.3
djangorestframework==3.15.1
django-filter==25.1
Pillow==9.0.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.0.0
django-environ==0.7.0
WeasyPrint      
twilio==8.0.0          
redis==5.0.0           
django-redis==5.3.0    
drf-spectacular==0.27.0
tabulate


